+++ src/Generator/Commands/AbstractCommand.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Rougin\Weasley\Commands;
+namespace Rougin\Weasley\Generator\Commands;
 
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;