+++ tests/Fixture/Templates/TestIntegration.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Skeleton\Integrations;
+namespace App\Integrations;
 
 use Rougin\Slytherin\Integration\Configuration;
 use Rougin\Slytherin\Container\ContainerInterface;
@@ -8,7 +8,7 @@ use Rougin\Slytherin\Container\ContainerInterface;
 /**
  * TestIntegration
  *
- * @package Skeleton
+ * @package App
  * @author  Rougin Royce Gutib <rougingutib@gmail.com>
  */
 class TestIntegration implements \Rougin\Slytherin\Integration\IntegrationInterface